+++ src/main/java/LTSaveEd/DataObjects/InventoryElements/Effect.java
@@ -1,4 +1,4 @@
-package LTSaveEd.Objects.InventoryElements;
+package LTSaveEd.DataObjects.InventoryElements;
 
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
