+++ src/main/java/LTSaveEd/DataObjects/InventoryElements/InventoryItem.java
@@ -1,6 +1,6 @@
-package LTSaveEd.Objects.InventoryElements;
+package LTSaveEd.DataObjects.InventoryElements;
 
-import LTSaveEd.Objects.InventoryElements.AbstractInventoryElements.AbstractInventoryElement;
+import LTSaveEd.DataObjects.InventoryElements.AbstractInventoryElements.AbstractInventoryElement;
 import org.w3c.dom.Node;
 
 public class InventoryItem extends AbstractInventoryElement {
