+++ src/main/java/LTSaveEd/DataObjects/LegTypeAttr.java
@@ -1,4 +1,4 @@
-package LTSaveEd.Objects;
+package LTSaveEd.DataObjects;
 
 import javafx.collections.ObservableList;
