+++ src/main/java/LTSaveEd/DataObjects/NpcCharacter.java
@@ -1,4 +1,4 @@
-package LTSaveEd.Objects;
+package LTSaveEd.DataObjects;
 
 import org.w3c.dom.NamedNodeMap;
 import org.w3c.dom.Node;
