+++ src/main/java/LTSaveEd/Util/InitializeElements.java
@@ -1,6 +1,6 @@
 package LTSaveEd.Util;
 
-import LTSaveEd.Objects.*;
+import LTSaveEd.DataObjects.*;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
